Add Get Org Users query
