+++ src/api/org.test.ts
@@ -1,5 +1,6 @@
 import { Observable } from 'rxjs';
 import { Api } from './api';
+import { RequestType } from '../constants';
 
 /**
  * Response
@@ -104,6 +105,84 @@ describe('Org Api', () => {
         let result = await api.org.get();
         expect(result).toThrow(TypeError);
       } catch (e) {}
+    });
+  });
+
+  /**
+   * Get Org Users
+   */
+  describe('Get Org Users', () => {
+    const response = {
+      status: 200,
+      statusText: 'OK',
+      ok: true,
+      data: [
+        {
+          id: 1,
+          name: '',
+          login: 'admin',
+          email: 'admin@localhost',
+          avatarUrl: '/avatar/46d229b033af06a191ff2267bca9ae56',
+          isAdmin: true,
+          isDisabled: false,
+          lastSeenAt: '2023-09-05T09:10:29Z',
+          lastSeenAtAge: '1 minute',
+          authLabels: ['label1', 'label2'],
+        },
+      ],
+      headers: {},
+      url: 'http://localhost:3000/api/datasources/proxy/1/api/users',
+      type: 'basic',
+      redirected: false,
+      config: {
+        method: 'GET',
+        url: 'api/datasources/proxy/1/api/users',
+        retry: 0,
+        headers: {
+          'X-Grafana-Org-Id': 1,
+        },
+        hideFromInspector: false,
+      },
+    };
+
+    const query = { refId: 'A', requestType: RequestType.GET_ORG_USERS };
+
+    it('Should make getUsers request', async () => {
+      fetchRequestMock = jest.fn().mockImplementation(() => getResponse(response));
+      let result = await api.org.getUsers();
+      expect(result).toBeTruthy();
+    });
+
+    it('Should not make getUsers request', async () => {
+      fetchRequestMock = jest.fn().mockImplementation(() => getResponse(undefined));
+
+      let result = await api.org.getUsers();
+      expect(result).toBeTruthy();
+      expect(result.length).toBe(0);
+    });
+
+    it('Should throw exception getUsers request', async () => {
+      fetchRequestMock = jest.fn().mockImplementation(() => getErrorResponse(response));
+
+      try {
+        let result = await api.org.getUsers();
+        expect(result).toThrow(TypeError);
+      } catch (e) {}
+    });
+
+    it('Should make getUsersFrame request', async () => {
+      fetchRequestMock = jest.fn().mockImplementation(() => getResponse(response));
+      let result = await api.org.getUsersFrame(query);
+      expect(result?.length).toEqual(1);
+      expect(result[0].fields.length).toEqual(10);
+    });
+
+    it('Should handle getUsersFrame request with no data', async () => {
+      fetchRequestMock = jest.fn().mockImplementation(() => getResponse(response));
+      response.data = [];
+
+      let result = await api.org.getUsersFrame(query);
+      expect(result?.length).toEqual(0);
     });
   });
 });
