+++ src/datasource/datasource.ts
@@ -65,8 +65,8 @@ export class DataSource extends DataSourceApi<Query, DataSourceOptions> {
           case RequestType.HEALTH:
             frames = await this.api.health.getFrame(target);
             break;
-          case RequestType.USERS:
-            frames = await this.api.users.getFrame(target);
+          case RequestType.GET_ORG_USERS:
+            frames = await this.api.org.getUsersFrame(target);
             break;
         }
